% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calcPropInteractionsGLVadjMat}
\alias{calcPropInteractionsGLVadjMat}
\title{Calculate the proportion of interacions types in a Lotka-Volterra adjacency matrix}
\usage{
calcPropInteractionsGLVadjMat(adjM, spc)
}
\arguments{
\item{adjM}{Numeric Matrix Lotka-Volterra adjcency matrix}

\item{sp}{Numeric vector with the present species, if species=0 the interaction is not counted}
}
\value{
Numeric vector with 5 elements Matrix 1=Mutualistic (+,+), 2=Predator-prey (-,+), 3=Competitive (-,-)
                  4=Comensalism (+,0), 5=Amensalism (-,0)
}
\description{
Calculate the proportion of interacions types in a Lotka-Volterra adjacency matrix
}
