% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateGLVparmsFromAdj}
\alias{generateGLVparmsFromAdj}
\title{Generate random parameters for Lotka-Volterra generalized interaction matrix from Lotka-Volterra adjacency matrix}
\usage{
generateGLVparmsFromAdj(adjM, ef, predIntAvg = 0.01,
  selfLimAvg = as.numeric(c(0.01, 0.01, 0.01)), migrAvg = 0,
  preserveInt = 0L)
}
\arguments{
\item{adjM}{NumericMatrix, metacommunity interaction matrix, the values and the signs are 
the parameters of the LV model then a predator prey interaction will be adjM(i,j)=-1 
adjM(j,i)=1 where j is the predator and i the prey. Competitive interaction is adjM(i,j)=adjM(j,i)=-1}

\item{ef}{Eficiency of predator prey interactions}

\item{predIntAvg}{Average value for the interaction intensity of predator-preys, competition, mutalistic.}

\item{selfLimAvg}{Numeric vector, and  also set the average value for diagonal entries of the interaction matrix
that represent self-limitation, the elements of the vector represent 1=mutualistic, 2=Basal, 3=predator species.}

\item{migrAvg}{Average value to generate at uniform random migration from [0,migrAvg*2]}

\item{preserveInt}{if 0 the values are random uniform with prdIntAvg as a mean, if 1 the values of the interactions adjM[i,i] when i!=j are preserved.
if 2 the values are random uniform with mean given by adjM[i,i] when i!=j.}
}
\value{
A list with the interaction matrix interM, the intrinsic growth rates r, and migration values m
}
\description{
Generate random parameters for Lotka-Volterra generalized interaction matrix from Lotka-Volterra adjacency matrix
}
