% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateGLVparms.R
\name{generateGLVparmsFromAdjR}
\alias{generateGLVparmsFromAdjR}
\title{R function to generate random parameters for the Lotka-Volterra model from an adjacency matrix.}
\usage{
generateGLVparmsFromAdjR(
  adjM,
  ef,
  predIntAvg = 0.01,
  selfLimAvg = c(0.01, 0.01, 0.01),
  migrAvg = 0,
  preserveInt = 0,
  predIntSd = 0.01,
  rndType = 0
)
}
\arguments{
\item{adjM}{Numeric matrix, metacommunity interaction matrix, the values and the signs are
the parameters of the LV model then a predator prey interaction will be adjM[i,j]=-1
adjM[j,i]=1 where j is the predator and i the prey. Competitive interaction is adjM[i,j]=adjM[j,i]=-1}

\item{ef}{Efficiency of predator prey interactions}

\item{predIntAvg}{Average value for the interaction intensity of predator-preys, competition, mutalistic.}

\item{selfLimAvg}{Numeric vector, and  also set the average value for diagonal entries of the interaction matrix
that represent self-limitation, the elements of the vector represent 1=mutualistic, 2=Basal, 3=predator species.}

\item{migrAvg}{Average value to generate a uniform random migration with range [0,migrAvg*2]}

\item{preserveInt}{if 0 the values are random defined by `rndType` with `predIntAvg` as a mean, if 1 the values of the interactions adjM[i,i] when i!=j are preserved.
if 2 the values are random uniform with mean given by adjM[i,i] when i!=j.}

\item{predIntSd}{Standard deviation of the distribution to generate the interaction values for `preserveInt = 0` and `rndType = 1` gamma distribution.
For uniform distribution the sd is fixed at `predIntAvg*2 / sqrt(12)`}

\item{rndType}{0=uniform, 1=gamma NOTE: this is implemented only for `preserveInt = 0`}
}
\value{
A list with the interaction matrix interM, the intrinsic growth rates r, and migration values m
}
\description{
There is an equivalent Rcpp function which presumably has memory issues.
}
