% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generateRandomGLVadjMat}
\alias{generateRandomGLVadjMat}
\title{Generate random Lotka-Volterra adjacency matrix with fixed proportion of interactions}
\usage{
generateRandomGLVadjMat(numSp, C, propInt)
}
\arguments{
\item{numSp}{Integer, number of species}

\item{C}{Double, Probability of interaction = Connectivity = number of interacions/(numSp^2)}

\item{propInt}{Numeric vector, Proportion of interactions, there are 5 types of interactions the elements of the vector
represent 1= mutualism (+,+), 2=Predator-prey (+,-), 3= competition (-,-), 4=Comensalims (+,0), 
5=Amensalism (-,0).}
}
\value{
A NumericMatrix with the GLV type of adjacency matrix
}
\description{
Generate random Lotka-Volterra adjacency matrix with fixed proportion of interactions
}
