% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetaWebAssemblyModels.R
\docType{package}
\name{MetaWebAssemblyModels}
\alias{MetaWebAssemblyModels}
\alias{MetaWebAssemblyModels-package}
\title{MetaWebAssemblyModels: A package for simulation of ecological network assembly}
\description{
The MetaWebAssemblyModels package provides stochastic models to simulate the assembly of ecological
networks from a metaweb or regional species pool.
}
