% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metaWebNetAssembly}
\alias{metaWebNetAssembly}
\title{Simulation of an Assembly process from a Meta-web assuming the interactions conserve in the local web}
\usage{
metaWebNetAssembly(metaW, m, e, time)
}
\arguments{
\item{metaW}{metacommunity adyacency matrix}

\item{m}{A numeric vector of species' migration rates (probability) from the meta-web}

\item{e}{A numeric vector of species' extinction probability}

\item{time}{Number of time steps of simulation}
}
\value{
A list with the final the number of species by time S, the number of links by time L, 
              the time series of species STime and the adjacency matrix A with effective links.
}
\description{
This is a dynamical model of colonization and extinction process, with the restriction 
that predators must have at least one prey species to survive in the local community,  described in [1]
}
\references{
1. Galiana, N., Lurgi, M., Claramunt-López, B., Fortin, M.-J., Leroux, S., Cazelles, K., et al. (2018). 
The spatial scaling of species interaction networks. Nat. Ecol. Evol., 2, 782–790
}
