% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{metaWebNetAssemblyCT}
\alias{metaWebNetAssemblyCT}
\title{Simulation of an Assembly process from a Meta-web assuming the interactions are conserved in the local web}
\usage{
metaWebNetAssemblyCT(metaW, m, e, se, time)
}
\arguments{
\item{metaW}{metacommunity adyacency matrix}

\item{m}{A numeric vector of species' migration rates (probability) from the meta-web}

\item{e}{A numeric vector of species' extinction probability}

\item{se}{A numeric vector of species' secondary extinction probability}

\item{time}{Number of time steps of simulation}
}
\value{
A list with the final the number of species by time S, the number of links by time L, 
              the time series of species STime and the adjacency matrix A with effective links.
}
\description{
This is a dynamical model of colonization and extinction process, similar to the TTIB model described in [1]. 
With an additional secondary extinctions probability that controls when a predator goes extinct if it has
no prey. This is a continuous time version of the model that follows the Gillespie algorithm [2] for simulation
There are 3 possible events migration with rate *m*, local extinction with rate *e*, and secondary extinction when
a predators has no prey with rate *se*. The parameters are vectors so each species could have their own parameter values.
}
\references{
1. Galiana, N., Lurgi, M., Claramunt-López, B., Fortin, M.-J., Leroux, S., Cazelles, K., et al. (2018). 
The spatial scaling of species interaction networks. Nat. Ecol. Evol., 2, 782–790
2. Gillespie, D. T. (1976). A general method for numerically simulating the stochastic time evolution of coupled chemical reactions. 
Journal of Computational Physics, 22(4), 403–434. doi: 10.1016/0021-9991(76)90041-3
}
