% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meweasmo.R
\docType{package}
\name{meweasmo}
\alias{meweasmo}
\title{meweasmo: A package for simulation of ecological network assembly}
\description{
The meweasmo package provides stochastic models to simulate the assembly of ecological
networks from a metaweb or regional species pool.
}
