% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{neutralNetAssembly}
\alias{neutralNetAssembly}
\title{Simulation of a neutral network assembly procesess}
\usage{
neutralNetAssembly(rho, m, q, a, time, BB = 0L)
}
\arguments{
\item{rho}{metacommunity number of species}

\item{m}{migration rate (probability) from the meta-web}

\item{q}{probability of interaction}

\item{a}{probability of extinction}

\item{time}{time of simulation}

\item{BB}{number of initial basal species present}
}
\description{
This model simulates an assembly process starting from basal species, that could be fixed with the 
  parameter BB or allowed to migrate with rate m. The interactions are established with probability q with 
  no restrictions, execpt that double interactions A->B, B->A and self interactions are not allowed. The model 
  simulates secondary extinctions, when a predator has all his preys extinct it will become locally exctinct.
}
